refactor: 依赖大版本升级 + 错误类型化 + SQL 下沉 + 构建性能优化
- deps: rmcp 1.3→2.2、rusqlite 0.32→0.40、r2d2_sqlite 0.25→0.35；
  MSRV 1.88→1.95（libsqlite3-sys 0.38 需 cfg_select!），CI 同步
- cc-server: handlers 全链路 CcResult；新增 CcError::{BuildBusy,
  InvalidParams,NotFound,IndexUnavailable}；MCP 出口统一映射
  -32602/-32603，瞬态错误携带 data.retryable；参数对象
  （ImpactArgs/TraceArgs/FlowArgs/ExploreOptions）
- cc-db: SQL 下沉——新增 FrameworkScanSession（单连接 prepare_cached
  扫描会话）；read_conn 收紧 pub(crate)；上层 crate rusqlite 仅剩
  dev-deps（测试种子）
- perf(scan): file-state 快照跨构建缓存（files_state 聚合 token）+
  watcher 定向扫描（BuildScope::Targeted），消增量构建 scan_diff
  的 O(repo) 地板
- perf(rss): 全量构建 prepare 阶段直写 staging 临时库
  （build_temp_db_staging/swap_rebuild_staging），PreparedBuild
  瘦身为文件清单 + epoch floor，消 O(repo) RSS 峰值
- docs: MSRV、MCP 错误契约、STORAGE/INDEXING/CONCURRENCY、
  TEST_PLAN 基线 1206+14、基准刷新（warm p95 全部 <500ms）

+++ crates/cc-db/src/index_db_query.rs
@@ -311,7 +311,7 @@ impl IndexDb {
             .map_err(db_err)?;
         let rows = stmt
             .query_map([], |row| {
-                Ok((row.get::<_, String>(0)?, row.get::<_, usize>(1)?))
+                Ok((row.get::<_, String>(0)?, row.get::<_, i64>(1)? as usize))
             })
             .map_err(db_err)?;
         rows.collect::<Result<Vec<_>, _>>().map_err(db_err)
